backend/requirements.txt
Django>=5.2,<6.0
psycopg2-binary>=2.9,<3.0
python-decouple>=3.8,<4.0
python-dotenv>=1.0,<2.0
whitenoise>=6.6,<7.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
requests>=2.31,<3.0
django-cors-headers>=4.3,<5.0
gunicorn>=21.2,<22.0
dj-database-url>=2.1,<3.0
django-jazzmin>=2.6,<3.0
Pillow>=10.0,<12.0
stripe>=7.0,<8.0
resend>=0.8.0,<1.0
python-dateutil>=2.8,<3.0
